backend/requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.40.59
botocore==1.40.59
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
greenlet==3.2.4
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
mypy-boto3-s3==1.40.26
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.44
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
